#2035: Share draft endpoints in single file while frontend migrates
